requirements.txt
flask
requests
python-dotenv
pytest
fireworks-ai